Catch IllegalStateException and throw CameraException
Should resolve the issue as seen below. Starting the preview can throw an IllegalStateException with no way to catch it. This catches the ISE and replaces it with the CameraException allowing you to catch the failure and process it rather than forcing a crash.

Fatal Exception: java.lang.IllegalStateException: Target display surface has not been set.
       at io.fotoapparat.hardware.v2.session.SessionProvider.getPreviewSession(SessionProvider.java:67)
       at io.fotoapparat.hardware.v2.session.SessionManager.startPreview(SessionManager.java:26)
       at io.fotoapparat.hardware.v2.Camera2.startPreview(Camera2.java:103)
       at io.fotoapparat.routine.StartCameraRoutine.tryToStartCamera(StartCameraRoutine.java:67)
       at io.fotoapparat.routine.StartCameraRoutine.run(StartCameraRoutine.java:47)
       at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1162)
       at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:636)
       at java.lang.Thread.run(Thread.java:764)